add successful login test
update the class name
