--- src/test/java/school/redrover/SaucedemoTest.java
@@ -1,36 +1,0 @@
-package school.redrover;
-
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.support.ui.ExpectedConditions;
-import org.openqa.selenium.support.ui.WebDriverWait;
-import org.testng.Assert;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
-import org.testng.annotations.Test;
-
-import java.time.Duration;
-
-public class SaucedemoTest {
-    private WebDriver driver;
-    private static final String BASE_URL = "https://www.saucedemo.com/";
-
-    @BeforeMethod
-    public void setUp() {
-        driver = new ChromeDriver();
-        driver.manage().window().maximize();
-        driver.get(BASE_URL);
-    }
-
-    @Test
-    public void testHomePageTitle() {
-        new WebDriverWait(driver, Duration.ofSeconds(5)).until(ExpectedConditions.urlToBe(BASE_URL));
-
-        Assert.assertEquals(driver.getTitle(), "Swag Labs");
-    }
-
-    @AfterMethod
-    public void tearDown() {
-        driver.quit();
-    }
-}
